src/qumo/transformer.py
"""
QUMO Transformer - Core code transformation and optimization engine.
"""

from typing import Any, Optional
import ast
import libcst as cst
from sympy import Symbol
import networkx as nx
from . import PATTERNS, QUANTUM_OPS, FLOW_PATTERNS

class QuantumTransformer(cst.CSTTransformer):
    """Transforms Python code using quantum-inspired patterns and Unicode optimization."""
    
    def __init__(self) -> None:
        super().__init__()
        self.ꜱ = Symbol('ψ')  # Quantum state symbol
        self.ᴘ = nx.DiGraph()  # Pattern dependency graph
        self.ᴡ = {}  # Symbol workspace
    
    def optimize_name(self, name: str) -> str:
        """Convert standard names to optimized Unicode patterns."""
        return PATTERNS.get(name, name)
    
    def apply_quantum_pattern(self, node: cst.CSTNode) -> cst.CSTNode:
        """Apply quantum-inspired code patterns."""
        match node:
            case cst.Name(value=v):
                return cst.Name(value=self.optimize_name(v))
            case cst.Call(func=f):
                # Transform function calls into quantum operations
                if isinstance(f, cst.Name):
                    op = QUANTUM_OPS.get(f.value)
                    if op:
                        return self.transform_to_quantum_op(node, op)
        return node
    
    def transform_to_quantum_op(self, node: cst.Call, op: str) -> cst.Call:
        """Transform a function call into its quantum operation equivalent."""
        match op:
            case "⟨·⟩":  # Expectation value
                return self.build_expectation_value(node)
            case "⊗":    # Tensor product
                return self.build_tensor_product(node)
            case _:      # Default transformation
                return node
    
    def build_expectation_value(self, node: cst.Call) -> cst.Call:
        """Build a quantum expectation value expression."""
        return cst.Call(
            func=cst.Name("expectation_value"),
            args=[
                cst.Arg(value=arg.value)
                for arg in node.args
            ]
        )
    
    def build_tensor_product(self, node: cst.Call) -> cst.Call:
        """Build a quantum tensor product expression."""
        return cst.Call(
            func=cst.Name("tensor_product"),
            args=[
                cst.Arg(value=arg.value)
                for arg in node.args
            ]
        )
    
    def visit_Name(self, node: cst.Name) -> cst.Name:
        """Visit and transform name nodes."""
        return self.apply_quantum_pattern(node)
    
    def visit_Call(self, node: cst.Call) -> cst.Call:
        """Visit and transform function calls."""
        return self.apply_quantum_pattern(node)

class FlowOptimizer:
    """Optimizes code flow using graph-based analysis and pattern matching."""
    
    def __init__(self) -> None:
        self.ꜱ = nx.DiGraph()  # Control flow graph
        self.ᴘ = {}  # Pattern cache
    
    def optimize_flow(self, ast_tree: ast.AST) -> ast.AST:
        """Optimize code flow using graph analysis."""
        self.build_flow_graph(ast_tree)
        return self.apply_flow_patterns(ast_tree)
    
    def build_flow_graph(self, node: ast.AST) -> None:
        """Build a control flow graph from AST."""
        for child in ast.walk(node):
            if isinstance(child, (ast.If, ast.While, ast.For)):
                self.ꜱ.add_edge(child, child.body[0])
    
    def apply_flow_patterns(self, node: ast.AST) -> ast.AST:
        """Apply flow optimization patterns."""
        # Implementation of flow pattern optimization
        return node

def optimize_code(source: str) -> str:
    """
    Main entry point for code optimization.
    
    Args:
        source: Source code to optimize
    
    Returns:
        Optimized source code with quantum patterns and Unicode optimization
    """
    # Parse source into CST
    source_tree = cst.parse_module(source)
    
    # Apply quantum transformations
    quantum_transformer = QuantumTransformer()
    transformed_tree = source_tree.visit(quantum_transformer)
    
    # Apply flow optimizations
    flow_optimizer = FlowOptimizer()
    ast_tree = ast.parse(transformed_tree.code)
    optimized_ast = flow_optimizer.optimize_flow(ast_tree)
    
    # Generate optimized code
    return ast.unparse(optimized_ast) 